requirements.txt
transformers==4.35.0
torch==2.1.0
accelerate==0.24.0
sentencepiece==0.1.99
huggingface-hub==0.19.0
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
loguru==0.7.2
requests==2.31.0
pinecone-client==2.2.4
langchain==0.1.0
langchain-community==0.0.10
langchain-openai==0.0.2
langchain-huggingface==0.0.6
sentence-transformers==2.2.2
numpy==1.24.3
pandas==2.0.3 